登录接口优化

+++ user-biz/src/main/java/com/jaagro/user/biz/mapper/CustomerUserMapper.java
@@ -6,37 +6,37 @@ import com.jaagro.user.biz.entity.CustomerUser;
 public interface CustomerUserMapper {
     /**
      *
-     * @mbggenerated 2018-08-18
+     * @mbggenerated 2018-08-23
      */
-    int deleteByPrimaryKey(Long id);
+    int deleteByPrimaryKey(Integer id);
 
     /**
      *
-     * @mbggenerated 2018-08-18
+     * @mbggenerated 2018-08-23
      */
     int insert(CustomerUser record);
 
     /**
      *
-     * @mbggenerated 2018-08-18
+     * @mbggenerated 2018-08-23
      */
     int insertSelective(CustomerUser record);
 
     /**
      *
-     * @mbggenerated 2018-08-18
+     * @mbggenerated 2018-08-23
      */
-    CustomerUser selectByPrimaryKey(Long id);
+    CustomerUser selectByPrimaryKey(Integer id);
 
     /**
      *
-     * @mbggenerated 2018-08-18
+     * @mbggenerated 2018-08-23
      */
     int updateByPrimaryKeySelective(CustomerUser record);
 
     /**
      *
-     * @mbggenerated 2018-08-18
+     * @mbggenerated 2018-08-23
      */
     int updateByPrimaryKey(CustomerUser record);
 
@@ -59,5 +59,5 @@ public interface CustomerUserMapper {
      * @param id
      * @return
      */
-    UserInfo getUserInfoById(Long id);
+    UserInfo getUserInfoById(Integer id);
 }
